Extract course list business logic to view model.
